Remove the log printing process. Make it app's responsibility.
The printing process was converted to a function that the application
should call from its process when Crystal is not active.

Some timing constants were redesigned. There is a common constant that
defines how much time Crystal should allocate for the application.

The pre-epoch app callback time is now configurable.